requirements.txt
setuptools
wheel
numpy
scipy
jsonschema
pypng
requests
boto3